Add a `reactor` world
Similar to how there exists a `command` world this commit adds a
`reactor` world which has everything that a `command` does except for
the export of the `run` interface. This mirrors the
`preview1-adapter-reactor` that Wasmtime currently has and is intended
to make this available for other WIT-based projects as well.

This commit additionally refactors the preexisting `command` world to
replace all of its `import`s with a single `include` of this new
`reactor` world.