feat!: Rename reactor to imports
We have established a convention for WASI proposals
to have a world definition called `imports`.

This change updates the `command` world to include
an `imports` world rather than `reactor`. The goal
is to establish consistency with this change as
well as clear up some confusion around command
vs reactors in this context.

Additionally, this change takes advantage of the `include`
syntax to include the `imports` worlds across WASI
proposals.